moved old root reducer into new file project reducer, created auth reducer and configured firestore reducer to be used in the root reducer
