+++ package.json
@@ -15,10 +15,12 @@
 		"coingecko-api": "^1.0.9",
 		"cryptocurrency-icons": "^0.16.1",
 		"date-fns": "^2.9.0",
+		"firebase": "^7.7.0",
 		"react": "^16.12.0",
 		"react-dom": "^16.12.0",
 		"react-ga": "^2.7.0",
 		"react-redux": "^7.1.3",
+		"react-redux-firebase": "^3.1.0",
 		"react-router": "^5.1.2",
 		"react-router-dom": "^5.1.2",
 		"react-scripts": "3.3.0",
