+++ src/store/reducers/authReducer.js
@@ -1,0 +1,8 @@
+const initialState = {};
+
+export const authReducer = (state = initialState, action) => {
+  switch (action.type) {
+    default:
+      return state;
+  }
+};
